USB: Swapping the USB part out - twas unobtanium
the 4UCONN connector that Adafruit used is basically made of unobtanium
and more or less doesn't exist.  The part that was trying to be used
the rear prongs did not match up by any stretch of the imagination.

A mess of e-mails an annoyance ensued.

I'm swapping out the part on everything I can now, this seems to be
more common, and there's parts that are similar enough to it that
replacing it won't be a big deal.  New part is the
Amphenol FCI 10118193-0001LF